Add clarifications on settings and functionality
Before it was not mentioned that Wagtail Headless Preview needs to be added 
to the list of installed apps or that it comes with migrations that need to be run. 
While this becomes clear when reading its docs, it is more convenient to include 
it here. 

Also, I was surprised that the tab that opens when I click on the "Preview" button is 
not opening actually the defined URL. Because I did not have an actual preview site 
in place at this time, I only got a default "page not found" from my browser. I think
the added explanation should should help other new users to not get stuck on that.