List Grapple URLs after Wagtail URLs
Before, the grapple URLs were loaded after the Wagtail URLs. 
This caused the Wagtail serving mechanism trying to serve the GraphQL endpoint 
and results in 403, because the serving mechanism does not expect POST requests. 

The Grapple URLs are now loaded before the Wagtail URLs. 
This enables the GraphQL endpoint to properly respond to requests.