Add test for image query
Image queries before have been tested through some page tests, but not
directly.

The new test is only minimal but could be extended for more image fields
that can be queried.

The test includes the `src` filed on the image query. This test serves as
an example to add the same test for the documents next and then implement
the filed on the document similar to how it exists on the image already.