Fix docker build (#210)
running `docker buiild .` in /example failed with an error that the
chained `../requirements.txt` could not be found. This commit moves
the Dockerfile to the root folder so the entire repo will be in the
default docker build context and includes the local wagtail-grapple
packaged in editable mode so the `example` site for any commit can be
easily published.