fix: Unlink of example/db.sqlite3 failed
The second instruction for developers is to remove this file. It creates
issues when switching branches during development with the server
running. This commit removes it. If standard test data is needed, the
project should use factories or fixtures instead of a backed in DB for
better cross platform compatibilty and consistency.