Replace SnippetObjectType with SnippetInterface (#405)
* Rename CustomInterface to AdditionalInterface

The name `CustomInterface` is too similar with `CustomPageInterface` and
the upcoming `CustomSnippetInterface` and their custom interfaces
settings file. Better rename it to `AdditionalInterface` so there is
less confusion.

* Rename the custom interface settings file

The settings file will contain an override for the default snippet
interface, in addition to the default page interface, so we're renaming
it up front here.

* Replace SnippetObjectType with SnippetInterface

Fixes https://github.com/torchbox/wagtail-grapple/issues/386

API clients could use a field on snippet objects to determine the type
of snippet they are looking at. Therefore, we change the snippet type to
an interface, similar to the page interface, so it can expose a new
field called `snippetType`.

* Update changelog
* Add contentType field on SnippetInterface
* Test for no registered snippets
* Create an instance each of both snippet types